install and configure Material UI
I chose to install MUI Icons and also all of MUI. I recall the
team enjoying the use of MUI as I'm growing to like it a lot.
There's also risk of it being overkill for this project if it
was to exist as-is and not grow. However, I could see this
scaling up quickly where MUI would be very helpful.